Moved java files
In order to build the java files the needs to be located in the java dir, otherwise gradle won't pick the up.

This can be verified by calling:
./gradlew clean install
jar tvf examples/build/install/examples/lib/examples-0.8-SNAPSHOT.jar | grep helloworld
