+++ examples/src/main/java/actor/proto/examples/java/helloworld/Hello.java
@@ -1,4 +1,4 @@
-package actor.proto.examples.helloworld_java;
+package actor.proto.examples.java.helloworld;
 
 import actor.proto.Props;
 
@@ -7,7 +7,7 @@ import java.io.IOException;
 import static actor.proto.Protos.PID;
 import static actor.proto.java.Actors.*;
 
-class hello {
+public class Hello {
     public static void main(String[] args) throws IOException {
         Props props = fromProducer(HelloActor::new);
         PID pid = spawn(props);
